Worgen_Zone_Phase_Fix
Phase is fixed from Duskhaven to Tempest Reach

PLEASE BE AWARE this is a phase fix it does not mean that every quest works (maybe one is not complete :P ), npc are fine and not doubled.

We are working on to bring you a Worgen Zone with the most PVE content and it is on the right way now.

Other fixes in this area are coming in next days.

Please be patient!

Co-Authored-By: monady <monady@users.noreply.github.com>